add LRU-LC code demo
