--- .travis.yml
@@ -1,7 +1,0 @@
-language: c
-sudo: enabled
-script: 
-  - git clone https://github.com/troydhanson/uthash.git
-  - cp uthash/src/uthash.h .
-  - make
-  - ./cardinality
