+++ ExactCardinality/dump1.dat
